+++ apps/cli/src/commands/whoami.ts
@@ -1,6 +1,6 @@
 import { type Printer, chalk } from "@codemod-com/printer";
 
-import { getCurrentUserData, getOrgsNames } from "#utils.js";
+import { getCurrentUserData, getOrgsNames } from "#auth-utils.js";
 
 export const handleWhoAmICommand = async (options: {
   printer: Printer;
